move former macos examples to common example app
