--- MacOSExample/src/config.tsx
@@ -1,1 +1,0 @@
-export const USE_NATIVE_DRIVER = true;
