+++ example/src/recipes/simpleRotate/index.tsx
