+++ example/src/recipes/simpleRotate/swmansion.png
