+++ example/src/simple/manual/index.tsx
